requirements.txt
torch
torchvision
gradio
opencv-python
pillow
pycocotools
numpy
